lua_preview/requirements.txt
lupa>=2.0
pygame-ce>=2.5
Pillow>=10.0
watchdog>=3.0
